Code reorganization
This commit improves code organization by extracting functionalities
from the `__init__.py` file and dividing them into more specialized
Python modules. In particular:

 - An `argparser.py` module contains the argument parser for the cwltest
   CLI
 - A `main.py` module contains the code to run the tests
 - A `__main__.py` file serves as the main entrypoint for the package
 - Some logic has been moved from the `__init__.py` file to the
   `utils.py` module
