+++ cwltest/__main__.py
@@ -1,0 +1,5 @@
+"""Default entrypoint for the cwltest module."""
+
+from . import main
+
+main.main()
